Sound Mechanics
Added sound mechanics to the game. Sound is a scene extending Area2D and containing a collider. SoundListener is a scene which listens if it enters the radius of a Sound and emits the Sound. This listener can be added to a scene, and the signal can be connected to a script.