Finetuned enemy movement
Made use of the lerp method and adjusted velocity calculations so that gravity would still apply. I'm using both the simple if not is_on_floor function for gravity as well as the area2d for it.

Timer has been implemented to update player's location when alarm is triggered, and rocks or other sounds DO NOT affect enemies when they spot the player.